fix: improves the app's button component
With the new dice roll feature, there was a need to improve the button
component to cover text buttons. All buttons used by the app are now
managed by a single component.

The icons were isolated in a summary file located in the icon
directory, which can be easily imported by any component.

This is one of the fixes generated by 528063e.