fix: replaces PNG images with SVG
The PNG images being used in the app icons have been replaced by SVG.
For this, some changes were made using React Native to support the
use of SVG.

An icon for the app, a new name, Devinte, and a splash screen image
were also added.

A button bug was found after implementing the propagation of actions.
The buttons with propagations are no longer supporting a single
action.